Move the wrapper class to a separated file
