+++ obs-studio-server/CMakeLists.txt
@@ -222,6 +222,9 @@ add_executable(
 	"${PROJECT_SOURCE_DIR}/source/nodeobs_settings.h"
 	"${PROJECT_SOURCE_DIR}/source/util-memory.cpp"
 	"${PROJECT_SOURCE_DIR}/source/util-memory.h"
+
+	###### wrapper-obs ######
+	"${PROJECT_SOURCE_DIR}/source/wrapper-obs.h"
 )
 
 if(WIN32)
